Add missing tests and documentation
The following changes is added:
 - Add `selectAll` method to the Handsontable.Core class;
 - Update TypeScript definition file;
 - Fix TypeScript tests which didn't work at all (the code wasn't
   compiled by tsc);
 - Add Handsontable alias to the TypeScript tests;
 - Move logic responsible to setting the selection through the mouse
   events to the separate file;
 - Extend `selectRows` and `selectColumns` methods for
   `keepPreviousSelection` option which can be use to making (or not)
non-contiguous selection;
 - Optimize `addClass` and `removeClass` helpers;
 - Add missing documentation;
 - Add missing tests;
 - Add several matchers to the jasmine which can be helpful for writing
   faster and easier expectations about rendered table selection.

Issue: #4875
