+++ test/e2e/core/selectRows.spec.js
@@ -10,91 +10,128 @@ describe('Core.selectRows', () => {
     }
   });
 
-  it('should mark single row visually (default selectionMode)', () => {
+  it('should mark single row visually (default selectionMode, without headers)', () => {
     handsontable({
       data: Handsontable.helper.createSpreadsheetObjectData(6, 4),
     });
 
     selectRows(2);
 
-    expect([
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-      ['A', '0', '0', '0'],
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-    ]).toBeMatchToSelectionPattern();
+    expect(`
+      |   :   :   :   |
+      |   :   :   :   |
+      | A : 0 : 0 : 0 |
+      |   :   :   :   |
+      |   :   :   :   |
+      |   :   :   :   |
+      `).toBeMatchToSelectionPattern();
+  });
+
+  it('should mark single row visually (default selectionMode)', () => {
+    handsontable({
+      data: Handsontable.helper.createSpreadsheetObjectData(6, 4),
+      colHeaders: true,
+      rowHeaders: true,
+    });
+
+    selectRows(2);
+
+    expect(`
+      |   ║ - : - : - : - |
+      |===:===:===:===:===|
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      | * ║ A : 0 : 0 : 0 |
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      `).toBeMatchToSelectionPattern();
   });
 
   it('should mark range of the rows visually (default selectionMode)', () => {
     handsontable({
       data: Handsontable.helper.createSpreadsheetObjectData(6, 4),
+      colHeaders: true,
+      rowHeaders: true,
     });
 
     selectRows(2, 3);
 
-    expect([
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-      ['A', '0', '0', '0'],
-      ['0', '0', '0', '0'],
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-    ]).toBeMatchToSelectionPattern();
+    expect(`
+      |   ║ - : - : - : - |
+      |===:===:===:===:===|
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      | * ║ A : 0 : 0 : 0 |
+      | * ║ 0 : 0 : 0 : 0 |
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      `).toBeMatchToSelectionPattern();
   });
 
   it('should mark range of the rows visually (default selectionMode, reversed selection)', () => {
     handsontable({
       data: Handsontable.helper.createSpreadsheetObjectData(6, 4),
+      colHeaders: true,
+      rowHeaders: true,
     });
 
     selectRows(3, 2);
 
-    expect([
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-      ['0', '0', '0', '0'],
-      ['A', '0', '0', '0'],
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-    ]).toBeMatchToSelectionPattern();
+    expect(`
+      |   ║ - : - : - : - |
+      |===:===:===:===:===|
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      | * ║ 0 : 0 : 0 : 0 |
+      | * ║ A : 0 : 0 : 0 |
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      `).toBeMatchToSelectionPattern();
   });
 
   it('should mark only single cell visually when selectionMode is set as `single', () => {
     handsontable({
       data: Handsontable.helper.createSpreadsheetObjectData(6, 4),
+      colHeaders: true,
+      rowHeaders: true,
       selectionMode: 'single',
     });
 
     selectRows(2);
 
-    expect([
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-      ['_', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-    ]).toBeMatchToSelectionPattern();
+    expect(`
+      |   ║ - :   :   :   |
+      |===:===:===:===:===|
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      | - ║ # :   :   :   |
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      `).toBeMatchToSelectionPattern();
   });
 
   it('should mark the range of the rows visually when selectionMode is set as `range`', () => {
     handsontable({
       data: Handsontable.helper.createSpreadsheetObjectData(6, 4),
+      colHeaders: true,
+      rowHeaders: true,
       selectionMode: 'range',
     });
 
     selectRows(1, 2);
 
-    expect([
-      [' ', ' ', ' ', ' '],
-      ['A', '0', '0', '0'],
-      ['0', '0', '0', '0'],
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-      [' ', ' ', ' ', ' '],
-    ]).toBeMatchToSelectionPattern();
+    expect(`
+      |   ║ - : - : - : - |
+      |===:===:===:===:===|
+      |   ║   :   :   :   |
+      | * ║ A : 0 : 0 : 0 |
+      | * ║ 0 : 0 : 0 : 0 |
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      |   ║   :   :   :   |
+      `).toBeMatchToSelectionPattern();
   });
 
   it('should not deselect current selection when selectRows is called without arguments', () => {
