+++ test/e2e/publicAPI.spec.js
@@ -152,4 +152,59 @@ describe('Public API', () => {
       expect(Handsontable.cellTypes.time.validator).toBe(Handsontable.validators.TimeValidator);
     });
   });
+
+  describe('Helpers', () => {
+    it('should expose all registered helpers', () => {
+      expect(Handsontable.dom.addClass).toBeFunction();
+      expect(Handsontable.dom.addEvent).toBeFunction();
+      expect(Handsontable.dom.closest).toBeFunction();
+      expect(Handsontable.dom.closestDown).toBeFunction();
+      expect(Handsontable.dom.empty).toBeFunction();
+      expect(Handsontable.dom.fastInnerHTML).toBeFunction();
+      expect(Handsontable.dom.fastInnerText).toBeFunction();
+      expect(Handsontable.dom.getCaretPosition).toBeFunction();
+      expect(Handsontable.dom.getComputedStyle).toBeFunction();
+      expect(Handsontable.dom.getCssTransform).toBeFunction();
+      expect(Handsontable.dom.getParent).toBeFunction();
+      expect(Handsontable.dom.getScrollLeft).toBeFunction();
+      expect(Handsontable.dom.getScrollTop).toBeFunction();
+      expect(Handsontable.dom.getScrollableElement).toBeFunction();
+      expect(Handsontable.dom.getScrollbarWidth).toBeFunction();
+      expect(Handsontable.dom.getSelectionEndPosition).toBeFunction();
+      expect(Handsontable.dom.getSelectionText).toBeFunction();
+      expect(Handsontable.dom.getStyle).toBeFunction();
+      expect(Handsontable.dom.getTrimmingContainer).toBeFunction();
+      expect(Handsontable.dom.getWindowScrollLeft).toBeFunction();
+      expect(Handsontable.dom.getWindowScrollTop).toBeFunction();
+      expect(Handsontable.dom.hasClass).toBeFunction();
+      expect(Handsontable.dom.hasHorizontalScrollbar).toBeFunction();
+      expect(Handsontable.dom.hasVerticalScrollbar).toBeFunction();
+      expect(Handsontable.dom.index).toBeFunction();
+      expect(Handsontable.dom.innerHeight).toBeFunction();
+      expect(Handsontable.dom.innerWidth).toBeFunction();
+      expect(Handsontable.dom.isChildOf).toBeFunction();
+      expect(Handsontable.dom.isChildOfWebComponentTable).toBeFunction();
+      expect(Handsontable.dom.isImmediatePropagationStopped).toBeFunction();
+      expect(Handsontable.dom.isInput).toBeFunction();
+      expect(Handsontable.dom.isLeftClick).toBeFunction();
+      expect(Handsontable.dom.isRightClick).toBeFunction();
+      expect(Handsontable.dom.isVisible).toBeFunction();
+      expect(Handsontable.dom.offset).toBeFunction();
+      expect(Handsontable.dom.outerHeight).toBeFunction();
+      expect(Handsontable.dom.outerWidth).toBeFunction();
+      expect(Handsontable.dom.overlayContainsElement).toBeFunction();
+      expect(Handsontable.dom.pageX).toBeFunction();
+      expect(Handsontable.dom.pageY).toBeFunction();
+      expect(Handsontable.dom.polymerUnwrap).toBeFunction();
+      expect(Handsontable.dom.polymerWrap).toBeFunction();
+      expect(Handsontable.dom.removeClass).toBeFunction();
+      expect(Handsontable.dom.removeEvent).toBeFunction();
+      expect(Handsontable.dom.removeTextNodes).toBeFunction();
+      expect(Handsontable.dom.resetCssTransform).toBeFunction();
+      expect(Handsontable.dom.setCaretPosition).toBeFunction();
+      expect(Handsontable.dom.setOverlayPosition).toBeFunction();
+      expect(Handsontable.dom.stopImmediatePropagation).toBeFunction();
+      expect(Handsontable.dom.stopPropagation).toBeFunction();
+    });
+  });
 });
