+++ test/types/index.types.ts
@@ -1,0 +1,6 @@
+import * as Handsontable from '../../handsontable';
+
+const baseVersion = Handsontable.baseVersion;
+const buildDate = Handsontable.buildDate;
+const packageName = Handsontable.packageName;
+const version = Handsontable.version;
