+++ test/types/methods.types.ts
@@ -1,0 +1,95 @@
+import * as Handsontable from 'handsontable';
+
+const elem = document.createElement('div');
+const hot = new Handsontable(elem, {});
+
+hot.addHook('foo', [() => {}]);
+hot.addHook('foo', () => {});
+hot.addHookOnce('foo', [() => {}]);
+hot.addHookOnce('foo', () => {});
+hot.alter('insert_row', 123, 123, 'foo', true);
+hot.alter('insert_row');
+hot.clear();
+hot.colOffset();
+hot.colToProp(123);
+hot.countCols();
+hot.countEmptyCols(true);
+hot.countEmptyRows(true);
+hot.countRenderedCols();
+hot.countRenderedRows();
+hot.countRows();
+hot.countSourceRows();
+hot.countVisibleCols();
+hot.countVisibleRows();
+hot.deselectCell();
+hot.destroy();
+hot.destroyEditor(true);
+hot.emptySelectedCells();
+hot.getActiveEditor();
+hot.getCell(123, 123, true);
+hot.getCellEditor(123, 123);
+hot.getCellMeta(123, 123);
+hot.getCellRenderer(123, 123);
+hot.getCellValidator(123, 123);
+hot.getColHeader(123);
+hot.getColWidth(123);
+hot.getCoords(elem.querySelector('td'));
+hot.getCopyableData(123, 123);
+hot.getCopyableText(123, 123, 123, 123);
+hot.getData(123, 123, 123, 123);
+hot.getDataAtCell(123, 123);
+hot.getDataAtCol(123);
+hot.getDataAtProp(123);
+hot.getDataAtRow(123);
+hot.getDataAtRowProp(123, 'foo');
+hot.getDataType(123, 123, 123, 123);
+hot.getInstance();
+hot.getPlugin('foo');
+hot.getRowHeader(123);
+hot.getRowHeight(123);
+hot.getSchema();
+hot.getSelected();
+hot.getSelectedLast();
+hot.getSelectedRange();
+hot.getSelectedRangeLast();
+hot.getSettings();
+hot.getSourceData(123, 123, 123, 123);
+hot.getSourceDataAtCell(123, 123);
+hot.getSourceDataAtCol(123);
+hot.getSourceDataAtRow(123);
+hot.getValue();
+hot.hasColHeaders();
+hot.hasHook('foo');
+hot.hasRowHeaders();
+hot.isEmptyCol(123);
+hot.isEmptyRow(123);
+hot.isListening();
+hot.listen();
+hot.loadData([]);
+hot.populateFromArray(123, 123, [], 123, 123, 'foo', 'foo', 'foo', []);
+hot.propToCol('foo');
+hot.propToCol(123);
+hot.removeCellMeta(123, 123, 'foo');
+hot.removeHook('foo', function() {});
+hot.render();
+hot.rowOffset();
+hot.runHooks('foo', 123, 'foo', true, {}, [], function() {});
+hot.selectAll();
+hot.selectCell(123, 123, 123, 123, true, true);
+hot.selectCellByProp(123, 'foo', 123, 'foo', true);
+hot.selectCells([[123, 'prop1', 123, 'prop2']], true, true);
+hot.selectCells([[123, 123, 123, 123]], true, true);
+hot.selectColumns(1, 4);
+hot.selectColumns(1);
+hot.selectRows(1, 4);
+hot.selectRows(1);
+hot.setCellMeta(123, 123, 'foo', 'foo');
+hot.setCellMetaObject(123, 123, {});
+hot.setDataAtCell(123, 123, 'foo', 'foo');
+hot.setDataAtCell(123, 123, {myProperty: 'foo'}, 'foo');
+hot.setDataAtRowProp(123, 'foo', 'foo', 'foo');
+hot.spliceCol(123, 123, 123, 'foo');
+hot.spliceRow(123, 123, 123, 'foo');
+hot.unlisten();
+hot.updateSettings({}, true);
+hot.validateCells(function() {});
